Break darling_core::error into multiple files
